+++ app/src/main/java/org/sopt/and/ui/component/textField/WavveCommonPasswordField.kt
@@ -1,7 +1,6 @@
-package org.sopt.and.ui.component
+package org.sopt.and.ui.component.textField
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -19,7 +18,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
@@ -33,16 +31,12 @@ import org.sopt.and.ui.theme.Gray5
 import org.sopt.and.ui.theme.WavveDisabled
 
 @Composable
-fun SignUpPasswordField(
+fun WavveCommonPasswordField(
     value: String,
     onValueChange: (String) -> Unit,
     hint: String,
-    isValid: Boolean,
 ) {
     var passwordVisible by remember { mutableStateOf(false) }
-    var isFocused by remember { mutableStateOf(false) }
-    val borderColor =
-        if (value.isNotEmpty() && !isFocused && !isValid) Color.Magenta else Color.Transparent
     val focusManager = LocalFocusManager.current
     val keyboardController = LocalSoftwareKeyboardController.current
 
@@ -87,10 +81,6 @@ fun SignUpPasswordField(
                 color = Gray5,
                 shape = RoundedCornerShape(7.dp),
             )
-            .border(1.dp, borderColor, RoundedCornerShape(7.dp))
             .height(48.dp)
-            .onFocusChanged {
-                isFocused = it.isFocused
-            }
     )
 }
